ajout dernier exo

+++ CPP06/ex01/Serializer.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Serializer.cpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: julietteandrieux <julietteandrieux@stud    +#+  +:+       +#+        */
+/*   By: juandrie <juandrie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/23 17:10:42 by julietteand       #+#    #+#             */
-/*   Updated: 2024/03/23 17:37:58 by julietteand      ###   ########.fr       */
+/*   Updated: 2024/03/25 11:27:01 by juandrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,12 +31,12 @@ Serializer::~Serializer()
 {
 }
 
-uintptr_t Serializer::serialize(Data *ptr)
+unsigned long Serializer::serialize(Data *ptr)
 {
-	return (reinterpret_cast<uintptr_t>(ptr));
+	return (reinterpret_cast<unsigned long>(ptr));
 }
 
-Data *Serializer::deserialize(uintptr_t raw)
+Data *Serializer::deserialize(unsigned long raw)
 {
 	return (reinterpret_cast<Data*>(raw));
 }
