+++ CPP06/ex01/main.cpp
@@ -3,12 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   main.cpp                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: julietteandrieux <julietteandrieux@stud    +#+  +:+       +#+        */
+/*   By: juandrie <juandrie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/23 17:16:27 by julietteand       #+#    #+#             */
-/*   Updated: 2024/03/23 17:57:57 by julietteand      ###   ########.fr       */
+/*   Created: 2024/03/25 11:25:29 by juandrie          #+#    #+#             */
+/*   Updated: 2024/03/25 11:25:30 by juandrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+
 
 #include "Serializer.hpp"
 
@@ -20,7 +22,7 @@ int main()
     originalData.text = "Example text";
 
     // Sérialise l'objet Data.
-    uintptr_t serialized = Serializer::serialize(&originalData);
+    unsigned long serialized = Serializer::serialize(&originalData);
     std::cout << "Serialized data: " << serialized << std::endl;
 
     // Désérialise l'objet Data.
@@ -35,7 +37,7 @@ int main()
    
 	// Test avec un décalage d'adresse pour vérifier la robustesse
     std::cout << "Testing with offset address:" << std::endl;
-    uintptr_t offsetSerialized = serialized + sizeof(int);
+    unsigned long offsetSerialized = serialized + sizeof(int);
     Data *offsetDeserializedData = Serializer::deserialize(offsetSerialized);
 
     // Cela devrait échouer car le pointeur désérialisé ne correspond pas à l'adresse d'origine.
@@ -46,7 +48,7 @@ int main()
 
     // Test de désérialisation d'un pointeur nul
     std::cout << "Testing with null pointer:" << std::endl;
-    uintptr_t nullSerialized = Serializer::serialize(NULL);
+    unsigned long nullSerialized = Serializer::serialize(NULL);
     Data* nullDeserializedData = Serializer::deserialize(nullSerialized);
 
     // Cela devrait réussir car un pointeur nul reste un pointeur nul après sérialisation/désérialisation.
